pyproject.toml
[project]
name = "datasette-events-forward"
version = "0.1a3"
description = "Forward Datasette events to another instance"
readme = "README.md"
authors = [{name = "Simon Willison"}]
license = "Apache-2.0"
classifiers=[
    "Framework :: Datasette",
]
requires-python = ">=3.10"
dependencies = [
    "datasette>=1.0a20",
    "python-ulid",
    "aiolimiter",
    "typing_extensions"
]

[project.urls]
Homepage = "https://github.com/datasette/datasette-events-forward"
Changelog = "https://github.com/datasette/datasette-events-forward/releases"
Issues = "https://github.com/datasette/datasette-events-forward/issues"
CI = "https://github.com/datasette/datasette-events-forward/actions"

[project.entry-points.datasette]
events_forward = "datasette_events_forward"

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "pytest-httpx", "datasette-test"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
